fix: local docker compose setup
Migrations are done automatically when starting the server and should not be applied in the local docker compose image itself

Revert "fix: local docker compose setup"

This reverts commit 96a82e3fe7b6fab9140788bef7eee9ded989f6ba.

a
