+++ src/server.ts
@@ -3,7 +3,6 @@ import { Server, IncomingMessage, ServerResponse } from 'node:http'
 import { runMigrations } from './utils/migrate'
 import { createServer } from './app'
 import { getConfig } from './utils/config'
-import { Client } from 'pg'
 import { logger } from './logger'
 
 const config = getConfig()
@@ -16,30 +15,10 @@ const main = async () => {
     requestIdHeader: 'Request-Id',
   })
 
-  // Init config
-  const port = getConfig().PORT
-
-  // Init DB
-  const dbConfig = {
-    connectionString: config.DATABASE_URL,
-    connectionTimeoutMillis: 10_000,
-  }
-  const client = new Client(dbConfig)
-
-  // Run migrations
-  try {
-    await client.connect()
-
-    // Ensure schema exists, not doing it via migration to not break current migration checksums
-    await client.query(`CREATE SCHEMA IF NOT EXISTS ${config.SCHEMA};`)
-
-    await runMigrations(client)
-  } finally {
-    await client.end()
-  }
+  await runMigrations()
 
   // Start the server
-  app.listen({ port: Number(port), host: '0.0.0.0' }, (err, address) => {
+  app.listen({ port: Number(config.PORT), host: '0.0.0.0' }, (err, address) => {
     if (err) {
       console.error(err)
       process.exit(1)
